redraw ui look, and redraw sidebar to be top layout;
