+++ xchange-frontend/src/App.jsx
@@ -1,7 +1,5 @@
 import { useState } from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
 import CurrencyConverterLayout from './Components/CurrencyConverterLayout';
