+++ xchange-frontend/src/Components/CurrencyConverter/CurrencyConverter.css
@@ -1,0 +1,79 @@
+:root {
+  --background-dark: #1c1c1e;
+  --input-background: #2c2c2e;
+  --gold: #d4af37;
+  --rose-gold: #b76e79;
+  --text-color: var(--gold);
+  --accent-color: var(--rose-gold);
+}
+
+.converter-container {
+  max-width: 100%;
+  width: 700px;
+  margin: 1.5rem auto;
+  padding: 1.5rem;
+  background-color: var(--background-dark);
+  border: 1px solid var(--accent-color);
+  border-radius: 10px;
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  color: var(--text-color);
+  box-shadow: 0 0 20px rgba(183, 110, 121, 0.3);
+}
+
+h2 {
+  text-align: center;
+  margin-bottom: 1.5rem;
+  color: var(--text-color);
+}
+
+.currency-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.form-item {
+  display: flex;
+  flex-direction: column;
+}
+
+label {
+  margin-bottom: 0.25rem;
+  font-weight: 600;
+  color: var(--text-color);
+}
+
+input,
+select {
+  padding: 0.5rem;
+  background-color: var(--input-background);
+  border: 1px solid var(--accent-color);
+  border-radius: 6px;
+  font-size: 1rem;
+  color: var(--text-color);
+  outline: none;
+}
+
+input[readonly] {
+  background-color: #3a3a3c;
+}
+
+input::placeholder {
+  color: #999;
+}
+
+button {
+  padding: 0.6rem;
+  font-size: 1rem;
+  background-color: var(--accent-color);
+  color: black;
+  border: none;
+  border-radius: 6px;
+  font-weight: bold;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+button:hover {
+  background-color: #d48a94;
+}
